requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
datadog-api-client==2.33.1
dj-database-url==2.3.0
Django==5.1.7
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-filter==25.1
django-prometheus==2.3.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
elastic-transport==8.17.1
elasticsearch==8.17.2
graylog==1.0.0
gunicorn==22.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
kavenegar==1.1.2
kombu==5.4.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.48
psutil==7.0.0
pycryptodome==3.20.0  # Modern replacement for pycrypto
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-graylog==0.3.0
python-json-logger==3.3.0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
requests==2.32.3
sentry-sdk==2.1.1
six==1.16.0
smsir-python==1.0.8
sqlparse==0.5.2
typing_extensions==4.13.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
SQLAlchemy==2.0.25  # For advanced database operations
psycopg2-binary==2.9.9
